Improve SeveritySelector.tsx with first draft about options
