+++ src/components/form/helpers/customSchemas.ts
@@ -4,7 +4,6 @@ import { z } from 'zod'
 
 import { isEmail } from '@/src/components/form/helpers/validators'
 import { KLEROS_LIST_TYPES_KEYS } from '@/types/kleros/types'
-import { Severity_Keys } from '@/types/utils'
 
 // Why we need these schemas?
 // https://github.com/iway1/react-ts-form#dealing-with-collisions
@@ -189,7 +188,10 @@ export const ChallengePeriodTypeSchema = createUniqueFieldSchema(
 )
 
 export const SeverityTypeSchema = createUniqueFieldSchema(
-  z.enum(Severity_Keys),
+  z.object({
+    amountOfJurors: z.number(),
+    challengeBounty: z.string(),
+  }),
   'SeverityTypeSchema',
 )
