+++ types/utils.ts
@@ -96,6 +96,7 @@ export enum Severity {
   'Normal' = 1,
   'Above average' = 3,
   'Heavy' = 5,
+  'Custom' = 99,
 }
 
 export const Severity_Keys = ['Normal', 'Above average', 'Heavy'] as const
